# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MESH API
  version: 0.3.0
  license: { name: UNLICENSED, identifier: LicenseRef-UNLICENSED }
  description: |
    Contract for the MESH student-collaboration modular monolith. This milestone exposes account,
    profile, self-declared skills, goals, interests, desired skills, weekly availability, GitHub public-repository evidence,
    discovery, connections, project groups, persistent messages, tasks, calendar events, and availability polls.
    GitHub evidence is a limited supporting signal, never proof of expertise. LinkedIn is not collected or scraped.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /discover/{targetUserId}/pass:
    put:
      operationId: passCandidate
      summary: Mark a student as passed so they are excluded from future Discover results
      parameters: [ { $ref: '#/components/parameters/TargetUserId' } ]
      responses:
        '204': { description: Pass recorded }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /discover/{targetUserId}/save:
    put:
      operationId: saveCandidate
      summary: Save a student for later without excluding them from Discover
      parameters: [ { $ref: '#/components/parameters/TargetUserId' } ]
      responses:
        '204': { description: Saved state recorded }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: removeSavedCandidate
      summary: Remove a student from the signed-in student's saved candidates
      parameters: [ { $ref: '#/components/parameters/TargetUserId' } ]
      responses:
        '204': { description: Saved state removed; no action is needed if the student was not saved }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /blocks/{targetUserId}:
    put:
      operationId: blockStudent
      summary: Block a student; hide the connection and prevent direct messages while the block is active
      parameters: [ { $ref: '#/components/parameters/TargetUserId' } ]
      responses:
        '204': { description: Student blocked }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: unblockStudent
      summary: Remove a block created by the signed-in student
      parameters: [ { $ref: '#/components/parameters/TargetUserId' } ]
      responses:
        '204': { description: Block removed; succeeds when no block exists }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /connection-requests:
    post:
      operationId: createConnectionRequest
      summary: Send or reopen a connection request
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SendConnectionRequest' } } }
      responses:
        '201': { description: Pending request, content: { application/json: { schema: { $ref: '#/components/schemas/ConnectionRequest' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /connection-requests/incoming:
    get:
      operationId: listIncomingConnectionRequests
      summary: List pending requests sent to the signed-in student
      responses:
        '200': { description: Pending incoming requests, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ConnectionRequest' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /connection-requests/{requestId}/accept:
    patch:
      operationId: acceptConnectionRequest
      summary: Accept an incoming pending request and create a connection
      parameters: [ { $ref: '#/components/parameters/RequestId' } ]
      responses:
        '200': { description: Connection created, content: { application/json: { schema: { $ref: '#/components/schemas/Connection' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /connection-requests/{requestId}/decline:
    patch:
      operationId: declineConnectionRequest
      summary: Decline an incoming pending request
      parameters: [ { $ref: '#/components/parameters/RequestId' } ]
      responses:
        '204': { description: Request declined }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /connection-requests/{requestId}:
    delete:
      operationId: cancelConnectionRequest
      summary: Cancel a pending request sent by the signed-in student
      parameters: [ { $ref: '#/components/parameters/RequestId' } ]
      responses:
        '204': { description: Request cancelled }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /connections:
    get:
      operationId: listConnections
      summary: List established connections for the signed-in student, excluding active blocks
      responses:
        '200': { description: Connections, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Connection' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /connections/{connectionId}/messages:
    get:
      operationId: listConnectionMessages
      summary: List recent private messages in chronological order; connected students only
      parameters:
        - $ref: '#/components/parameters/ConnectionId'
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200': { description: Connection messages, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/ConnectionMessage' } } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: sendConnectionMessage
      summary: Send a private message; connected students only
      parameters: [ { $ref: '#/components/parameters/ConnectionId' } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SendConnectionMessageRequest' } } }
      responses:
        '201': { description: Message sent, content: { application/json: { schema: { $ref: '#/components/schemas/ConnectionMessage' } } } }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /github/authorization-url:
    post:
      operationId: beginGitHubAuthorization
      summary: Create a single-use GitHub OAuth authorization URL for public repository evidence
      description: Requests no GitHub scopes and uses PKCE. The frontend redirects the student to `authorizationUrl`; access tokens are never returned or persisted by MESH.
      responses:
        '200': { description: Authorization URL created, content: { application/json: { schema: { $ref: '#/components/schemas/GitHubAuthorizationUrl' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /github/callback:
    get:
      operationId: completeGitHubAuthorization
      summary: GitHub OAuth redirect callback
      description: Public browser callback only. On success or failure, redirects to the configured frontend URL. It is not called by the frontend API client.
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema: { type: string }
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: error
          in: query
          required: false
          schema: { type: string }
      responses:
        '302': { description: Redirected to configured success or error frontend URL }
  /github/connection:
    get:
      operationId: getGitHubConnection
      summary: Get GitHub public-evidence connection state for the signed-in student
      responses:
        '200': { description: Connection status, content: { application/json: { schema: { $ref: '#/components/schemas/GitHubConnection' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: disconnectGitHub
      summary: Delete collected GitHub repository evidence and disconnect GitHub
      responses:
        '204': { description: GitHub connection and stored public evidence deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /discord/authorization-url:
    post:
      operationId: beginDiscordAuthorization
      summary: Create a single-use Discord OAuth authorization URL to link a MESH account to a Discord identity
      description: Requests only Discord identity. Discord access tokens are exchanged server-side and never persisted.
      responses:
        '200': { description: Authorization URL created, content: { application/json: { schema: { $ref: '#/components/schemas/DiscordAuthorizationUrl' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /discord/callback:
    get:
      operationId: completeDiscordAuthorization
      summary: Discord OAuth redirect callback
      security: []
      parameters:
        - name: code
          in: query
          required: false
          schema: { type: string }
        - name: state
          in: query
          required: false
          schema: { type: string }
        - name: error
          in: query
          required: false
          schema: { type: string }
      responses:
        '302': { description: Redirected to configured success or error frontend URL }
  /discord/connection:
    get:
      operationId: getDiscordConnection
      summary: Get the signed-in student's linked Discord identity
      responses:
        '200': { description: Discord connection state, content: { application/json: { schema: { $ref: '#/components/schemas/DiscordConnection' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
    delete:
      operationId: disconnectDiscord
      summary: Remove the signed-in student's linked Discord identity
      responses:
        '204': { description: Discord identity removed }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/discord-room:
    get:
      operationId: getProjectDiscordRoom
      summary: Get the Discord room for a project; project members only
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Current room state, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectDiscordRoom' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createOrSyncProjectDiscordRoom
      summary: Create or synchronize a private Discord room; project owner only
      description: Every project member must link Discord before this endpoint grants room access.
      parameters: [ { $ref: '#/components/parameters/ProjectId' } ]
      responses:
        '200': { description: Created or synchronized room, content: { application/json: { schema: { $ref: '#/components/schemas/ProjectDiscordRoom' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    ValidationError:
      description: Request field validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Unauthorized:
      description: Missing, invalid, or expired bearer token
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflict:
      description: A unique value is already in use
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: The signed-in student is not allowed to perform this action
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: The requested resource does not exist or is not accessible
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ServiceUnavailable:
      description: The requested integration is not configured
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  parameters:
    TargetUserId:
      name: targetUserId
      in: path
      required: true
      schema: { type: string, format: uuid }
    RequestId:
      name: requestId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ConnectionId:
      name: connectionId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    GitHubAuthorizationUrl:
      type: object
      required: [authorizationUrl, expiresAt]
      properties:
        authorizationUrl: { type: string, format: uri, description: Open in the browser to let the student authorize collection of public repository evidence. }
        expiresAt: { type: string, format: date-time, description: Authorization URL expiration; currently 10 minutes after creation. }
    GitHubConnection:
      type: object
      required: [connected, publicRepositoryCount]
      properties:
        connected: { type: boolean }
        login: { type: [string, 'null'], description: Public GitHub login when connected. }
        avatarUrl: { type: [string, 'null'], format: uri }
        publicRepositoryCount: { type: integer, minimum: 0, description: "Number of most-recent owned public repositories stored as evidence, capped by server configuration." }
        authorizedAt: { type: [string, 'null'], format: date-time }
        lastSyncedAt: { type: [string, 'null'], format: date-time }
    DiscordAuthorizationUrl:
      type: object
      required: [authorizationUrl, expiresAt]
      properties:
        authorizationUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
    DiscordConnection:
      type: object
      required: [connected]
      properties:
        connected: { type: boolean }
        discordUserId: { type: [string, 'null'] }
        username: { type: [string, 'null'] }
        globalName: { type: [string, 'null'] }
        connectedAt: { type: [string, 'null'], format: date-time }
    ProjectDiscordRoom:
      type: object
      required: [available]
      properties:
        available: { type: boolean }
        channelId: { type: [string, 'null'] }
        inviteUrl: { type: [string, 'null'], format: uri }
        lastSyncedAt: { type: [string, 'null'], format: date-time }
...
    SendConnectionRequest:
      type: object
      required: [recipientUserId]
      properties:
        recipientUserId: { type: string, format: uuid }
    ConnectionRequest:
      type: object
      required: [id, senderId, senderUsername, senderName, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        senderUsername: { type: string }
        senderName: { type: string }
        status: { type: string, enum: [PENDING, ACCEPTED, DECLINED, CANCELLED] }
        createdAt: { type: string, format: date-time }
    Connection:
      type: object
      required: [id, collaborator, createdAt]
      properties:
        id: { type: string, format: uuid }
        collaborator: { $ref: '#/components/schemas/Collaborator' }
        createdAt: { type: string, format: date-time }
    SendConnectionMessageRequest:
      type: object
      required: [content]
      properties:
        content: { type: string, minLength: 1, maxLength: 2000 }
    ConnectionMessage:
      type: object
      required: [id, connectionId, senderId, senderName, content, createdAt]
      properties:
        id: { type: string, format: uuid }
        connectionId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        senderName: { type: string }
        content: { type: string }
        createdAt: { type: string, format: date-time }
    Collaborator:
      type: object
      required: [id, username, displayName]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        department: { type: [string, 'null'] }
        yearOfStudy: { type: [integer, 'null'] }
        bio: { type: [string, 'null'] }
        avatarKey: { type: [string, 'null'] }
...
    Error:
      type: object
      required: [timestamp, status, error, message, fields, path]
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
        fields: { type: object, additionalProperties: { type: string } }
        path: { type: string }